优化1

+++ modest-maven/src/main/java/cn/org/expect/maven/concurrent/MavenSearchJob.java
@@ -62,7 +62,9 @@ public abstract class MavenSearchJob extends BaseJob implements Runnable, MavenS
         } catch (Throwable e) {
             log.error(e.getLocalizedMessage(), e);
         } finally {
-            this.service.removeJob(this);
+            if (this.service != null) {
+                this.service.removeJob(this);
+            }
         }
     }
 }
